requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coverage==7.8.2
cryptography==45.0.3
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
limits==5.2.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
bcrypt==4.1.2
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.1.1
pytest-mock==3.14.1
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wrapt==1.17.2
